requirements.txt
streamlit
google_generativeai
python-dotenv
langchain_core
langchain_community
langchain_google_genai
PyPDF2
chromadb>=0.4.23
faiss-cpu